soc/power9/xbus.*: extract XBus helpers to a unit
Same kind of SCOM accesses are used since SBE of the second CPU has
started and until it can be accessed via XSCOM.  There are also a couple
of XBus-related constants.
