+++ src/include/cpu/power/istep_8.h
@@ -13,9 +13,4 @@ void istep_8_9(uint8_t chips);
 void istep_8_10(uint8_t chips);
 void istep_8_11(uint8_t chips);
 
-/* These functions access SCOM of the second CPU using SBE IO, thus they can be
- * used only in isteps that come after 8.4 */
-void put_scom(uint8_t chip, uint64_t addr, uint64_t data);
-uint64_t get_scom(uint8_t chip, uint64_t addr);
-
 #endif /* CPU_PPC64_ISTEP8_H */
