+++ src/soc/ibm/power9/Makefile.inc
@@ -38,6 +38,7 @@ romstage-y += mcbist.c
 romstage-y += timer.c
 romstage-y += fsi.c
 romstage-y += sbeio.c
+romstage-y += xbus.c
 ramstage-y += chip.c
 ramstage-y += homer.c
 ramstage-y += rom_media.c
